+++ dxlink-javascript/dxlink-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @dxfeed/dxlink-core
 
+## 0.9.0
+
 ## 0.8.1
 
 ### Patch Changes